+++ hdr/compiler_error.h
@@ -3,18 +3,18 @@
 #ifndef COMPILER_ERROR_H
 #define COMPILER_ERROR_H
 
-void lex_error(char c, int i);
+void lex_error(char c, long i);
 void parse_error(token *t);
-void typecheck_error(char *msg, int i);
+void typecheck_error(char *msg, long i);
 void ir_error(char *msg);
 void set_src(char *src);
 void set_verbose(bool mode);
 
 typedef struct {
-  int line;
-  int i;
+  long line;
+  long i;
 } src_location;
 
-void print_src_location(int i);
+void print_src_location(long i);
 
 #endif
