+++ hdr/lexer.h
@@ -4,12 +4,12 @@
 #define LEXER_H
 
 vector *lex(const char *src);
-int lex_str(const char *src, int i, token *t);
-int lex_wrd(const char *src, int i, token *t);
-int lex_num(const char *src, int i, token *t);
-int lex_nl(const char *src, int i, token *t);
-int lex_ws(const char *src, int i, token *t);
-int lex_op(const char *src, int i, token *t);
-int lex_pnct(const char *src, int i, token *t);
+long lex_str(const char *src, long i, token *t);
+long lex_wrd(const char *src, long i, token *t);
+long lex_num(const char *src, long i, token *t);
+long lex_nl(const char *src, long i, token *t);
+long lex_ws(const char *src, long i, token *t);
+long lex_op(const char *src, long i, token *t);
+long lex_pnct(const char *src, long i, token *t);
 
 #endif
