+++ hdr/parse_stmt.h
@@ -3,6 +3,6 @@
 #ifndef PARSE_STMT_H
 #define PARSE_STMT_H
 
-int parse_stmt(vector *tokens, int i, stmt *s);
+long parse_stmt(vector *tokens, long i, stmt *s);
 
 #endif
