+++ hdr/parse_type.h
@@ -4,7 +4,7 @@
 #ifndef PARSE_TYPE_H
 #define PARSE_TYPE_H
 
-int parse_type(vector *tokens, int i, type *t);
-int parse_type_arr(vector *tokens, int i, type *t);
+long parse_type(vector *tokens, long i, type *t);
+long parse_type_arr(vector *tokens, long i, type *t);
 
 #endif
