+++ hdr/safe.h
@@ -1,7 +1,7 @@
 #ifndef SAFE_H
 #define SAFE_H
 char *safe_strcat(char *lhs, const char *rhs);
-void *safe_alloc(int size);
-char *safe_realloc_str(char *ptr, int size);
+void *safe_alloc(long size);
+char *safe_realloc_str(char *ptr, long size);
 char *safe_strcpy(char *dptr, char *sptr);
 #endif
