+++ src/ir/asm_ir_print.c
@@ -8,14 +8,14 @@
 char *asm_prog_to_str(asm_prog *prog) {
   char *result = safe_alloc(1);
 
-  // Print globals and externs
+  // Prlong globals and externs
   result = safe_strcat(result, ASM_GLOBALS);
   result = safe_strcat(result, ASM_EXTERNS);
   result = safe_strcat(result, "\n");
 
   // Data section
   result = safe_strcat(result, "section .data\n");
-  for (int i = 0; i < prog->const_names->size; i++) {
+  for (long i = 0; i < prog->const_names->size; i++) {
     char *cur_name = vector_get_str(prog->const_names, i);
     char *cur_val = vector_get_str(prog->const_vals, i);
 
@@ -28,7 +28,7 @@ char *asm_prog_to_str(asm_prog *prog) {
 
   // Text section
   result = safe_strcat(result, "section .text\n");
-  for (int i = 1; i < prog->fns->size; i++) {
+  for (long i = 1; i < prog->fns->size; i++) {
     asm_fn *cur_fn = vector_get_asm_fn(prog->fns, i);
 
     result = safe_strcat(result, cur_fn->name);
@@ -38,7 +38,7 @@ char *asm_prog_to_str(asm_prog *prog) {
     result = safe_strcat(result, cur_fn->name);
     result = safe_strcat(result, ":\n");
 
-    for (int j = 0; j < cur_fn->code->size; j++) {
+    for (long j = 0; j < cur_fn->code->size; j++) {
       char *cur_segment = vector_get_str(cur_fn->code, j);
       result = safe_strcat(result, cur_segment);
     }
@@ -54,7 +54,7 @@ char *asm_prog_to_str(asm_prog *prog) {
   result = safe_strcat(result, jpl_main->name);
   result = safe_strcat(result, ":\n");
 
-  for (int j = 0; j < jpl_main->code->size; j++) {
+  for (long j = 0; j < jpl_main->code->size; j++) {
     char *cur_segment = vector_get_str(jpl_main->code, j);
     result = safe_strcat(result, cur_segment);
   }
