+++ src/parser/parse_stmt.c
@@ -1,13 +1,13 @@
 #include "parse_stmt.h"
-#include "safe.h"
 #include "compiler_error.h"
 #include "parse_expr.h"
 #include "parse_lval.h"
 #include "parser.h"
+#include "safe.h"
 #include "vector_get.h"
 #include <string.h>
 
-int parse_stmt(vector *tokens, int i, stmt *s) {
+long parse_stmt(vector *tokens, long i, stmt *s) {
   s->start = vector_get_token(tokens, i)->start;
 
   switch (peek_token(tokens, i)) {
